Added Analytics API filter, created script to run NZ_Script 1 and 2
Added API script
