+++ .gitignore
@@ -328,6 +328,7 @@ paket-files/
 # Python Tools for Visual Studio (PTVS)
 __pycache__/
 *.pyc
+venv/
 
 # Cake - Uncomment if you are using it
 # tools/**
