+++ CEFbibprocess.csv
@@ -1,0 +1,37 @@
+JobID,Network Id,Existing 035a,Incoming 035a,Action
+CEF_1012688_1748084545832::bib::1050196,991012558889705153,61605867,61605867,match
+CEF_1012688_1748084545832::bib::1145988,991011955409705153,63178558,63178558,match
+CEF_1012688_1748084545832::bib::1151031,991011918989705153,15866118,15866118,match
+CEF_1012688_1748084545832::bib::115204,991023419587205153,1154857448,1154857448,match
+CEF_1012688_1748084545832::bib::1154105,991011895369705153,74555622,74555622,match
+CEF_1012688_1748084545832::bib::133346,991023419586605153,1441717292,1441717292,match
+CEF_1012688_1748084545832::bib::1357435,991010581979705153,938461,938461,match
+CEF_1012688_1748084545832::bib::13883,991023421189005153,1453518967,1453518967,match
+CEF_1012688_1748084545832::bib::1527340,991009174139705153,462297948,462297948,match
+CEF_1012688_1748084545832::bib::1612022,991008475989705153,9469065,9469065,match
+CEF_1012688_1748084545832::bib::1614986,991008470539705153,3013204,3013204,match
+CEF_1012688_1748084545832::bib::1618812,991008421599705153,1043995879,1043995879,match
+CEF_1012688_1748084545832::bib::1620321,991008405239705153,1043988282,1043988282,match
+CEF_1012688_1748084545832::bib::1621964,991008402169705153,1044028830,1044028830,match
+CEF_1012688_1748084545832::bib::1625966,991008324949705153,294051,294051,match
+CEF_1012688_1748084545832::bib::178230,991023404387705153,1466951963,1466951963,match
+CEF_1012688_1748084545832::bib::1840540,991007387449705153,33367919,33367919,match
+CEF_1012688_1748084545832::bib::187058,991023393872305153,3741109,3741109,field transfer
+CEF_1012688_1748084545832::bib::1931763,991006536569705153,6526120,6526120,match
+CEF_1012688_1748084545832::bib::1996033,991005963529705153,1044015871,1044015871,match
+CEF_1012688_1748084545832::bib::204511,991023300834505153,762897842,762897842,match
+CEF_1012688_1748084545832::bib::209402,991023295333205153,731328334,731328334,match
+CEF_1012688_1748084545832::bib::210874,991023295333005153,42574457,42574457,match
+CEF_1012688_1748084545832::bib::215074,991023236733905153,731328337,731328337,match
+CEF_1012688_1748084545832::bib::218607,991023236732405153,694059159,694059159,match
+CEF_1012688_1748084545832::bib::221108,991023216825505153,846851107,846851107,match
+CEF_1012688_1748084545832::bib::22390,991023421188805153,1272864578,1272864578,match
+CEF_1012688_1748084545832::bib::244643,991023157626005153,1042080358,1042080358,match
+CEF_1012688_1748084545832::bib::25411,991023421188705153,1453195224,1453195224,match
+CEF_1012688_1748084545832::bib::2661773,991002269209705153,37132430,37132430,match
+CEF_1012688_1748084545832::bib::28296,991023421188605153,1112424450,1112424450,match
+CEF_1012688_1748084545832::bib::30429,991023421188505153,1482188793,1482188793,match
+CEF_1012688_1748084545832::bib::458223,991016172749705153,29311201,29311201,match
+CEF_1012688_1748084545832::bib::695318,991014377549705153,311984047,311984047,match
+CEF_1012688_1748084545832::bib::76811,991023419589605153,1460456396,1460456396,match
+CEF_1012688_1748084545832::bib::963147,991012985109705153,220016672,220016672,match
