+++ UWObibprocess.csv
@@ -1,0 +1,11 @@
+JobID,Network Id,Existing 035a,Incoming 035a,Action
+UWO_1012743_1748081412178::bib::1115452,991024699019705163,2129528,2129528,match
+UWO_1012743_1748081412178::bib::1551826,991014885289705163,1105288974,1105288974,match
+UWO_1012743_1748081412178::bib::1686403,991012511629705163,1066903087,1066903087,match
+UWO_1012743_1748081412178::bib::1712459,991012391859705163,1071821397,1071821397,match
+UWO_1012743_1748081412178::bib::2101386,991009643679705163,,1520914156,unresolved
+UWO_1012743_1748081412178::bib::556562,991045895616905163,1245231707,1245231707,match
+UWO_1012743_1748081412178::bib::576527,991045828414005163,,1372277189,match
+UWO_1012743_1748081412178::bib::62384,991046325962305163,1313398538,1313398538,match
+UWO_1012743_1748081412178::bib::702509,991045019878805163,1223011899,1223011899,match
+UWO_1012743_1748081412178::bib::903411,991038270649705163,61695988,61695988,match
