+++ api.py
@@ -1,0 +1,5 @@
+import os
+
+def get_api_key():
+    return os.getenv('NZ_API_KEY')
+
